Shut down existing streams if they exist
This leaves the stream structures intact, but puts them into "standby",
which essentially means that just the pcm is closed. The pcm is only
opened in the functions out_write and in_read, so in those functions,
we simply return before it opens the pcm with a return value that
indicates that the read or write was successful.

Change-Id: I0222f111ccab4dadad58fa4c89c7d731a1bbd975