Add error checking to dmhd1000 library
This fixes up a lot (most. all?) of the crashiness.
There is more work to be done. The entire library is junk, working
with ascii representations of bytes in hex. Really weird.

There is also a problem with reading from the serial port, which
seems to be most prevalent after the tuner has been closed and
reopened.

Change-Id: Idba02640f1a7725f0d85045ae35d679aed4ed1d1